lilygo-sim7000/utils/requirements.txt
google.cloud
google-cloud-bigquery
